Move the GetBrokerConfig function to Broker
