+++ pkg/resources/kafka/configmap.go
@@ -254,8 +254,25 @@ func generateListenerSpecificConfig(l *v1beta1.ListenersConfig, log logr.Logger)
 func (r Reconciler) generateBrokerConfig(id int32, brokerConfig *v1beta1.BrokerConfig, extListenerStatuses,
 	intListenerStatuses, controllerIntListenerStatuses map[string]v1beta1.ListenerStatusList,
 	serverPass, clientPass string, superUsers []string, log logr.Logger) string {
+	finalBrokerConfig := getFinalBrokerConfig(id, r.KafkaCluster, log)
+
+	// Get operator generated configuration
+	opGenConf := r.getConfigProperties(brokerConfig, id, extListenerStatuses, intListenerStatuses, controllerIntListenerStatuses, serverPass, clientPass, superUsers, log)
+
+	// Merge operator generated configuration to the final one
+	if opGenConf != nil {
+		finalBrokerConfig.Merge(opGenConf)
+	}
+
+	finalBrokerConfig.Sort()
+
+	return finalBrokerConfig.String()
+}
+
+// TODO move this into pkg/sdk in the future (adamantal)
+func getFinalBrokerConfig(id int32, kafkaCluster *v1beta1.KafkaCluster, log logr.Logger) *properties.Properties {
 	// Parse cluster-wide readonly configuration
-	finalBrokerConfig, err := properties.NewFromString(r.KafkaCluster.Spec.ReadOnlyConfig)
+	finalBrokerConfig, err := properties.NewFromString(kafkaCluster.Spec.ReadOnlyConfig)
 	if err != nil {
 		log.Error(err, "failed to parse readonly cluster configuration")
 	}
@@ -263,7 +280,7 @@ func (r Reconciler) generateBrokerConfig(id int32, brokerConfig *v1beta1.BrokerC
 	// Parse readonly broker configuration
 	var parsedReadOnlyBrokerConfig *properties.Properties
 	// Find configuration for broker with id
-	for _, broker := range r.KafkaCluster.Spec.Brokers {
+	for _, broker := range kafkaCluster.Spec.Brokers {
 		if broker.Id == id {
 			parsedReadOnlyBrokerConfig, err = properties.NewFromString(broker.ReadOnlyConfig)
 			if err != nil {
@@ -278,15 +295,5 @@ func (r Reconciler) generateBrokerConfig(id int32, brokerConfig *v1beta1.BrokerC
 		finalBrokerConfig.Merge(parsedReadOnlyBrokerConfig)
 	}
 
-	// Get operator generated configuration
-	opGenConf := r.getConfigProperties(brokerConfig, id, extListenerStatuses, intListenerStatuses, controllerIntListenerStatuses, serverPass, clientPass, superUsers, log)
-
-	// Merge operator generated configuration to the final one
-	if opGenConf != nil {
-		finalBrokerConfig.Merge(opGenConf)
-	}
-
-	finalBrokerConfig.Sort()
-
-	return finalBrokerConfig.String()
+	return finalBrokerConfig
 }
