+++ pkg/resources/kafka/pod_test.go
@@ -63,18 +63,22 @@ func TestGetAffinity(t *testing.T) {
 	affinity := getAffinity(&nilAffinityBrokerConfig, &cluster)
 	assert.DeepEqual(t, affinity.PodAntiAffinity, defaultPodAntiAffinity.PodAntiAffinity)
 
-	mergedAffinityBrokerConfig, _ := v1beta1.GetBrokerConfig(v1beta1.Broker{
+	broker := v1beta1.Broker{
 		BrokerConfig: &nilAffinityBrokerConfig,
-	}, cluster.Spec)
+	}
+
+	mergedAffinityBrokerConfig, _ := broker.GetBrokerConfig(cluster.Spec.BrokerConfigGroups)
 
 	// expecting old behavior
 	affinity = getAffinity(mergedAffinityBrokerConfig, &cluster)
 	assert.DeepEqual(t, affinity.PodAntiAffinity, defaultPodAntiAffinity.PodAntiAffinity)
 
-	mergedAffinityBrokerConfig2, _ := v1beta1.GetBrokerConfig(v1beta1.Broker{
+	broker = v1beta1.Broker{
 		BrokerConfigGroup: "test",
 		BrokerConfig:      &nilAffinityBrokerConfig,
-	}, cluster.Spec)
+	}
+
+	mergedAffinityBrokerConfig2, _ := broker.GetBrokerConfig(cluster.Spec.BrokerConfigGroups)
 
 	// expecting old behavior
 	affinity = getAffinity(mergedAffinityBrokerConfig2, &cluster)
